airflow/dags/Scripts/Pipeline_Scripts/requirements.txt
pydantic
beautifulsoup4
lxml
pandas
python-dotenv
snowflake-connector-python
snowflake-connector-python[pandas]
pathlib